Autocomplete: restructure text processing utils
