+++ vscode/src/completions/getInlineCompletions.ts
@@ -8,9 +8,9 @@ import { logError } from '../log'
 
 import { GetContextOptions, GetContextResult } from './context/context'
 import { DocumentHistory } from './context/history'
+import { detectMultiline } from './detect-multiline'
 import { DocumentContext } from './get-current-doc-context'
 import * as CompletionLogger from './logger'
-import { detectMultiline } from './multiline'
 import { CompletionProviderTracer, Provider, ProviderConfig, ProviderOptions } from './providers/provider'
 import { RequestManager, RequestParams } from './request-manager'
 import { reuseLastCandidate } from './reuse-last-candidate'
