+++ vscode/src/completions/text-processing/index.ts
@@ -1,0 +1,1 @@
+export * from './utils'
