+++ vscode/src/completions/text-processing/process-inline-completions.test.ts
@@ -1,17 +1,13 @@
 import { beforeAll, describe, expect, test } from 'vitest'
 import Parser from 'web-tree-sitter'
 
-import { range } from '../testutils/textDocument'
+import { range } from '../../testutils/textDocument'
+import { getCurrentDocContext } from '../get-current-doc-context'
+import { documentAndPosition, initTreeSitterParser } from '../testHelpers'
+import { updateParseTreeCache } from '../tree-sitter/parse-tree-cache'
+import { InlineCompletionItem } from '../types'
 
-import { getCurrentDocContext } from './get-current-doc-context'
-import {
-    addParseInfoToCompletions,
-    adjustRangeToOverwriteOverlappingCharacters,
-    processItem,
-} from './processInlineCompletions'
-import { documentAndPosition, initTreeSitterParser } from './testHelpers'
-import { updateParseTreeCache } from './tree-sitter/parse-tree-cache'
-import { InlineCompletionItem } from './types'
+import { adjustRangeToOverwriteOverlappingCharacters, processItem } from './process-inline-completions'
 
 describe('adjustRangeToOverwriteOverlappingCharacters', () => {
     test('no adjustment at end of line', () => {
@@ -20,7 +16,7 @@ describe('adjustRangeToOverwriteOverlappingCharacters', () => {
         expect(
             adjustRangeToOverwriteOverlappingCharacters(item, {
                 position,
-                docContext: { currentLineSuffix: '' },
+                currentLineSuffix: '',
             })
         ).toEqual<InlineCompletionItem>(item)
     })
@@ -31,7 +27,7 @@ describe('adjustRangeToOverwriteOverlappingCharacters', () => {
         expect(
             adjustRangeToOverwriteOverlappingCharacters(item, {
                 position,
-                docContext: { currentLineSuffix: ')' },
+                currentLineSuffix: ')',
             })
         ).toEqual<InlineCompletionItem>({
             ...item,
@@ -45,7 +41,7 @@ describe('adjustRangeToOverwriteOverlappingCharacters', () => {
         expect(
             adjustRangeToOverwriteOverlappingCharacters(item, {
                 position,
-                docContext: { currentLineSuffix: ') ' },
+                currentLineSuffix: ') ',
             })
         ).toEqual<InlineCompletionItem>({
             ...item,
@@ -54,7 +50,7 @@ describe('adjustRangeToOverwriteOverlappingCharacters', () => {
     })
 })
 
-describe('addParseInfoToCompletions', () => {
+describe('parseCompletion', () => {
     let parser: Parser
 
     beforeAll(async () => {
@@ -64,18 +60,11 @@ describe('addParseInfoToCompletions', () => {
     function testParseInfoProcessor(code: string, completioSnippets: string[]) {
         const { document, position } = documentAndPosition(code)
         const docContext = getCurrentDocContext(document, position, Infinity, Infinity)
-
-        const completions = completioSnippets.map(insertText =>
-            processItem({ insertText }, { document, position, multiline: insertText.includes('\n'), docContext })
-        )
-
         updateParseTreeCache(document, parser)
 
-        return addParseInfoToCompletions(completions, {
-            document,
-            position,
-            docContext,
-        })
+        return completioSnippets.map(insertText =>
+            processItem({ insertText }, { document, position, multiline: insertText.includes('\n'), docContext })
+        )
     }
 
     test('adds parse info to single-line completions', () => {
@@ -90,16 +79,16 @@ describe('addParseInfoToCompletions', () => {
         expect(completions.map(c => c.hasParseErrors)).toEqual([false, true])
     })
 
-    test('adds parse info to multi-line completions', () => {
+    test.only('adds parse info to multi-line completions', () => {
         const completions = testParseInfoProcessor(
             `
             function hello() {
                 alert('hello world!')
             }
 
-            function sort(█)
+            const wow = []; function sort(█)
         `,
-            ['array) {\nreturn array.sort()\n}', 'array) new\n']
+            ['array) {\nreturn array.sort()\n} function kek() {}', 'array) new\n']
         )
 
         expect(completions.map(c => c.hasParseErrors)).toEqual([false, true])
