+++ vscode/src/completions/text-processing/string-comparator.ts
